Adapt Roland's analysis script to the project.
The scripts Roland uses to analyze his benchmark runs produce very nice
reports, so I wanted to incorporate them into this repository.  I also
wanted to simplify it so that no external libraries are needed and one
call is enough to generate a PDF report.  Accomplishing all this meant
rewriting most of the scripts, but that should be done now.