Try to make tracer logging more robust.
Before I was using subprocess pipes to get stdout and stderr.  A problem
with this approach is that nothing gets written for jobs that run too
and are killed by SGE.  I tried to get around this by not piping the
output and simply reading the log files at the end of the simulation.
Unfortunately this didn't seem to solve the problem.