Generate logs even for jobs that die.
This was difficult because subprocess was caching stdout and stderr, so
if rosetta crashed for some reason (e.g. segfault or time limit) no
output was being produced.  The solution was to add a utilities.tee()
function that prints output as it becomes available.  I had to use
select() to handle stdout and stderr separately.