Add the ability to resume benchmarks.
You can now easily run more jobs for any benchmark.  There are a few
caveats, namely that the new jobs just use the same files as the old
jobs and won't notice if those files have changed in the meantime, and
that the right version of rosetta has to be checked out.

To facilitate this feature, loop_benchmark.py now gets all of the
information it needs (except a benchmark id) from the database instead
of the command line.  This simplifies the code, because it obviates the
need to copy options from kickoff.py to loop_benchmark.py.