Record the version of rosetta each benchmark uses.
Two new columns are added to the benchmarks table: git_commit and
git_diff.  The former if simply the currently checked out git commit
hash.  The latter is a diff showing any unstaged changes to files that
are already part of the repository.  These two fields should make it
possible to exactly identify which version of rosetta was used (if the
compilation step wasn't skipped.)

Note that the --resume option will still gladly resume a benchmark with
a different version of rosetta.