Store stdout and stderr as LONGTEXT.
Some jobs were generating enough output to overflow the max size of the
TEXT type.  Furthermore, this was causing the whole transaction to fail,
resulting in the loss of the benchmark->protocol id mapping.  I tweaked
the transaction to commit this important mapping information as soon as
possible, to mitigate the effects of future bugs like this.