Moving files.
Adding pruned versions of the original RCSB files and the Rosetta-minimized versions. These pruned PDB files are missing the loop residue atoms (N-terminus N and CA are kept as are the C-terminus CA and C atoms) and all sidechains with a heavy atom within 10 angstroms of a heavy atom of a loop residue. This pruning is intented to remove bias towards the native structure and represent the type of problem faced in applications where the loop coordinates are missing and the surrounding sidechains have only been partially determined.

The next step is to add versions of the pruned Rosetta structures with the missing residues added back in but not in the correct coordinates. This will allow us to run existing protocols (CCD, KIC, NGK) on these new structures.